+++ packages/subgraph/tests/token-voting/token-voting.test.ts
@@ -37,7 +37,6 @@ import {
   createTokenVotingProposalEntityState
 } from './utils';
 
-let proposalId = '0';
 let actions = createDummyActions(DAO_TOKEN_ADDRESS, '0', '0x00000000');
 
 test('Run TokenVoting (handleProposalCreated) mappings with mock event', () => {
@@ -51,7 +50,7 @@ test('Run TokenVoting (handleProposalCreated) mappings with mock event', () => {
   getProposalCountCall(CONTRACT_ADDRESS, '1');
   createGetProposalCall(
     CONTRACT_ADDRESS,
-    proposalId,
+    PROPOSAL_ID,
     true,
     false,
 
@@ -72,7 +71,7 @@ test('Run TokenVoting (handleProposalCreated) mappings with mock event', () => {
 
   // create event
   let event = createNewProposalCreatedEvent(
-    proposalId,
+    PROPOSAL_ID,
     ADDRESS_ONE,
     STRING_DATA,
     CONTRACT_ADDRESS
@@ -81,68 +80,63 @@ test('Run TokenVoting (handleProposalCreated) mappings with mock event', () => {
   // handle event
   _handleProposalCreated(event, DAO_ADDRESS, STRING_DATA);
 
-  let entityID =
-    Address.fromString(CONTRACT_ADDRESS).toHexString() +
-    '_' +
-    BigInt.fromString(proposalId).toHexString();
   let packageId = Address.fromString(CONTRACT_ADDRESS).toHexString();
 
   // checks
-  assert.fieldEquals('TokenVotingProposal', entityID, 'id', entityID);
-  assert.fieldEquals('TokenVotingProposal', entityID, 'dao', DAO_ADDRESS);
-  assert.fieldEquals('TokenVotingProposal', entityID, 'plugin', packageId);
-  assert.fieldEquals('TokenVotingProposal', entityID, 'proposalId', proposalId);
-  assert.fieldEquals('TokenVotingProposal', entityID, 'creator', ADDRESS_ONE);
-  assert.fieldEquals('TokenVotingProposal', entityID, 'metadata', STRING_DATA);
+  assert.fieldEquals('TokenVotingProposal', PROPOSAL_ID, 'id', PROPOSAL_ID);
+  assert.fieldEquals('TokenVotingProposal', PROPOSAL_ID, 'dao', DAO_ADDRESS);
+  assert.fieldEquals('TokenVotingProposal', PROPOSAL_ID, 'plugin', packageId);
+  assert.fieldEquals('TokenVotingProposal', PROPOSAL_ID, 'creator', ADDRESS_ONE);
+  assert.fieldEquals('TokenVotingProposal', PROPOSAL_ID, 'metadata', STRING_DATA);
   assert.fieldEquals(
     'TokenVotingProposal',
-    entityID,
+    PROPOSAL_ID,
     'createdAt',
     event.block.timestamp.toString()
   );
   assert.fieldEquals(
     'TokenVotingProposal',
-    entityID,
+    PROPOSAL_ID,
     'creationBlockNumber',
     event.block.number.toString()
   );
-  assert.fieldEquals('TokenVotingProposal', entityID, 'startDate', START_DATE);
+  assert.fieldEquals('TokenVotingProposal', PROPOSAL_ID, 'startDate', START_DATE);
 
   assert.fieldEquals(
     'TokenVotingProposal',
-    entityID,
+    PROPOSAL_ID,
     'votingMode',
     VOTING_MODES.get(parseInt(VOTING_MODE))
   );
   assert.fieldEquals(
     'TokenVotingProposal',
-    entityID,
+    PROPOSAL_ID,
     'supportThreshold',
     SUPPORT_THRESHOLD
   );
   assert.fieldEquals(
     'TokenVotingProposal',
-    entityID,
+    PROPOSAL_ID,
     'minParticipation',
     MIN_PARTICIPATION
   );
 
-  assert.fieldEquals('TokenVotingProposal', entityID, 'startDate', START_DATE);
-  assert.fieldEquals('TokenVotingProposal', entityID, 'endDate', END_DATE);
+  assert.fieldEquals('TokenVotingProposal', PROPOSAL_ID, 'startDate', START_DATE);
+  assert.fieldEquals('TokenVotingProposal', PROPOSAL_ID, 'endDate', END_DATE);
   assert.fieldEquals(
     'TokenVotingProposal',
-    entityID,
+    PROPOSAL_ID,
     'snapshotBlock',
     SNAPSHOT_BLOCK
   );
 
   assert.fieldEquals(
     'TokenVotingProposal',
-    entityID,
+    PROPOSAL_ID,
     'totalVotingPower',
     TOTAL_VOTING_POWER
   );
-  assert.fieldEquals('TokenVotingProposal', entityID, 'executed', 'false');
+  assert.fieldEquals('TokenVotingProposal', PROPOSAL_ID, 'executed', 'false');
 
   // check TokenVotingPlugin
   assert.fieldEquals(
@@ -328,12 +322,8 @@ test('Run TokenVoting (handleVoteCast) mappings with mock event and vote option 
 });
 
 test('Run TokenVoting (handleProposalExecuted) mappings with mock event', () => {
-  // create state
-  let entityID =
-    Address.fromString(CONTRACT_ADDRESS).toHexString() + '_' + '0x0';
-
   createTokenVotingProposalEntityState(
-    entityID,
+    PROPOSAL_ID,
     DAO_ADDRESS,
     CONTRACT_ADDRESS,
     ADDRESS_ONE
@@ -342,7 +332,7 @@ test('Run TokenVoting (handleProposalExecuted) mappings with mock event', () => 
   // create calls
   createGetProposalCall(
     CONTRACT_ADDRESS,
-    proposalId,
+    PROPOSAL_ID,
     true,
     true,
 
@@ -362,23 +352,23 @@ test('Run TokenVoting (handleProposalExecuted) mappings with mock event', () => 
   );
 
   // create event
-  let event = createNewProposalExecutedEvent('0', CONTRACT_ADDRESS);
+  let event = createNewProposalExecutedEvent(PROPOSAL_ID, CONTRACT_ADDRESS);
 
   // handle event
   handleProposalExecuted(event);
 
   // checks
-  assert.fieldEquals('TokenVotingProposal', entityID, 'id', entityID);
-  assert.fieldEquals('TokenVotingProposal', entityID, 'executed', 'true');
+  assert.fieldEquals('TokenVotingProposal', PROPOSAL_ID, 'id', PROPOSAL_ID);
+  assert.fieldEquals('TokenVotingProposal', PROPOSAL_ID, 'executed', 'true');
   assert.fieldEquals(
     'TokenVotingProposal',
-    entityID,
+    PROPOSAL_ID,
     'executionDate',
     event.block.timestamp.toString()
   );
   assert.fieldEquals(
     'TokenVotingProposal',
-    entityID,
+    PROPOSAL_ID,
     'executionBlockNumber',
     event.block.number.toString()
   );